Rename 'diagnostics' folder to 'logPoints'
