+++ src/logPoints/LogPointsDebugAdapter.ts
