+++ src/logPoints/LogPointsManager.ts
