+++ src/logPoints/logPointsClient.ts
