+++ src/logPoints/structs/CommandRunResult.ts
