+++ src/logPoints/structs/IAttachProcessRequest.ts
