+++ src/logPoints/structs/ICloseSessionRequest.ts
