+++ src/logPoints/structs/ICloseSessionResponse.ts
