+++ src/logPoints/structs/IDebugSessionMetaData.ts
