+++ src/logPoints/structs/IEnumerateProcessResponse.ts
