+++ src/logPoints/structs/IGetLogpointsRequest.ts
