+++ src/logPoints/structs/IGetLogpointsResponse.ts
