+++ src/logPoints/structs/ILoadedScriptsResponse.ts
