+++ src/logPoints/structs/IRemoveLogpointRequest.ts
