+++ src/logPoints/structs/IRemoveLogpointResponse.ts
