+++ src/logPoints/structs/ISetLogpointRequest.ts
