+++ src/logPoints/structs/ISetLogpointResponse.ts
