+++ src/logPoints/structs/IStartSessionRequest.ts
