+++ src/logPoints/structs/IStartSessionResponse.ts
