+++ src/logPoints/structs/Logpoint.ts
