+++ src/logPoints/structs/LogpointsCollection.ts
